@@ -1,7 +1,7 @@
 describe CASino::ServiceTicketsController do
   describe 'GET "validate"' do
     let(:request_options) { params.merge(use_route: :casino) }
-    let(:service_ticket) { FactoryGirl.create :service_ticket }
+    let(:service_ticket) { FactoryBot.create :service_ticket }
     let(:service) { service_ticket.service }
     let(:parameters) { { service: service, ticket: service_ticket.ticket }}
     let(:params) { parameters }
