@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'nokogiri'
 
 describe CASino::ServiceTicket::SingleSignOutNotifier do
-  let(:service_ticket) { FactoryGirl.create :service_ticket }
+  let(:service_ticket) { FactoryBot.create :service_ticket }
   let(:service) { service_ticket.service }
   let(:notifier) { described_class.new service_ticket }
 
